refactor: trim comments and split provider.rs into a module
Housecleaning pass over provider, provider_http, nostr and cashu.

Comments: dropped multi-paragraph module headers, "Unit N of the 12-month
plan" references, restatements of the next line, and essays about paths not
taken. Kept the non-obvious whys: the whitelist-before-network ordering,
why the HTTP path must never call the mint, the heartbeat d-tag bucketing,
the temp-file-and-rename in the workload mirror, and the delete-anyway
reasoning in the cleanup sweep.

provider.rs is now src/provider/{mod,config,handlers,persistence,standby}.rs.
Public paths are unchanged via re-exports. handle_spawn_request lost its
11-argument signature to a HandlerDeps struct and shed the container-config,
standby-reservation and access-details blocks into helpers.

Dead code: NostrRelaySubscriber::{publish_offer, convert_event},
OfferEventContent, send_provisioning_request_private_message,
derive_seed_from_nostr_key, ProviderStats::uptime_start.

Also: box PrivateRequest::Spawn (272-byte variant spread), truncate
unparseable DM content by chars rather than bytes so a multi-byte codepoint
at offset 100 can't panic the handler, and clear the remaining clippy
warnings in these files (needless clones, useless format!, wildcard arm,
manual Default, unwrap_or_else(Default::default)).
